Extra file for snow3g
